+++ src/mariner/trackers/linuxtracker.py
@@ -14,7 +14,7 @@ class LinuxTracker(trackerplugin.TrackerPlugin):
     log = logging.getLogger(__name__)
 
     search_url = "https://linuxtracker.org/index.php?page=torrents&search={title}"
-    filters = "legal"
+    filters = {"legal"}
 
     def _parse(self, raw: str) -> Iterator[torrent.Torrent]:  # pylint: disable=too-many-locals
         """Parse result page.