+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - New --anime/-A flag to search only on anime trackers.
+- New --filter/-F flag to search on trackers with certain tags. Enables general filters on trackers.
 
 ## [1.3.0] - 2019-01-14